Move scripts + add missing namespace
