--- Assets/_Core/Scripts/Inputs/ActivateDesactivateUI.cs
@@ -1,15 +1,0 @@
-using MoonlitMixes.Inputs;
-using UnityEngine;
-
-public class ActivateDesactivateUI : MonoBehaviour
-{
-    private void OnEnable()
-    {
-        InputManager.Instance.SwitchActionMap("UI");
-    }
-
-    private void OnDisable()
-    {
-        InputManager.Instance.SwitchActionMap("Player");
-    }
-}
