+++ Assets/_Core/Scripts/Players/PlayerInteractionShop.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 91bcffa169a27f4429132a3f2c07aaac
